extract remove button
