+++ extension/src/panel/app-routes/edit-route/style.module.css
@@ -6,18 +6,6 @@ h2 {
   margin: 0em;
 }
 
-button.removeButton {
-  height: auto;
-  width: auto;
-  aspect-ratio: 1/1;
-  padding: 4px;
-  border: 3px double crimson;
-}
-
-button.removeButton:hover {
-  background-color: rgba(220, 20, 60, 0.2);
-}
-
 .backLink {
   position: absolute;
   top: -25px;
